GLAN-187: BC Unit Testing
[1] BC Controller, first pass unit test
[2] Update dashboard for strong parameters in BC and BC Templates
controllers
[3] Update a few JS files to allow for strong parameters
[4] Update for BC list AJAX call, better return structure

+++ app/controllers/biomedical_concepts_controller.rb
@@ -27,67 +27,64 @@ class BiomedicalConceptsController < ApplicationController
     @bcs = BiomedicalConcept.list
     respond_to do |format|
       format.json do
-        results = {}
-        results[:aaData] = []
-        @bcs.each do |bc|
-          item = {:id => bc.id, :namespace => bc.namespace, :identifier => bc.identifier, :label => bc.label}
-          results[:aaData] << item
-        end
+        results = {:data => []}
+        @bcs.each { |x| results[:data] << x.to_json }
         render json: results
       end
     end
   end
-  
+
   def history
     authorize BiomedicalConcept
-    @identifier = params[:identifier]
-    @bc = BiomedicalConcept.history(params)
+    @identifier = the_params[:identifier]
+    @bc = BiomedicalConcept.history(the_params)
     redirect_to biomedical_concepts_path if @bc.count == 0
   end
 
-  def new_template
+  def new_from_template
     authorize BiomedicalConcept, :new?
-    uri = params[:uri]
-    parts = uri.split('#')
-    ns = parts[0]
-    id = parts[1]
-    @bct = BiomedicalConceptTemplate.find(id, ns)
+    uri = UriV2.new({uri: the_params[:uri]})
+    ConsoleLogger.debug(C_CLASS_NAME, "new_from_template", "URI=#{uri.to_s}")
+    @bct = BiomedicalConceptTemplate.find(uri.id, uri.namespace)
+  end
+
+  def create
+    authorize BiomedicalConcept
+    @bc = BiomedicalConcept.create_simple(the_params)
+    if @bc.errors.empty?
+      flash[:success] = 'Biomedical Concept was successfully created.'
+      AuditTrail.create_item_event(current_user, @bc, "Biomedical Concept created.")
+      redirect_to biomedical_concepts_path
+    else
+      flash[:error] = @bc.errors.full_messages.to_sentence
+      redirect_to biomedical_concepts_path
+    end
   end
 
   def edit
     authorize BiomedicalConcept
-    ns = params[:namespace]
-    id = params[:id]
-    @bc = BiomedicalConcept.find(id, ns)
+    @bc = BiomedicalConcept.find(params[:id], the_params[:namespace])
   end
 
   def clone
     authorize BiomedicalConcept
-    ns = params[:namespace]
-    id = params[:id]
-    @bc = BiomedicalConcept.find(id, ns)
+    @bc = BiomedicalConcept.find(params[:id], the_params[:namespace])
   end
 
-  #def impact
-  #  authorize BiomedicalConcept, :view?
-  #  id = params[:id]
-  #  namespace = params[:namespace]
-  #  @bc = BiomedicalConcept.find(id, namespace)
-  #  @forms = Form.bc_impact(params)
-  #  @domains = SdtmUserDomain.bc_impact(params)
-  #end
-
-  def create
-    authorize BiomedicalConcept
-    instance = params[:instance]
-    @bc = BiomedicalConcept.create(params)
+  def clone_create
+    authorize BiomedicalConcept, :create?
+    @bc = BiomedicalConcept.create_clone(the_params)
     if @bc.errors.empty?
-      render :json => { :instance => instance, :data => @bc.to_edit}, :status => 200
+      AuditTrail.create_item_event(current_user, @bc, "BiomedicalConcept cloned from #{identifier}.")
+      flash[:success] = 'Biomedical Concept was successfully created.'
+      redirect_to forms_path
     else
-      render :json => { :errors => @bc.errors.full_messages}, :status => 422
+      flash[:error] = @bc.errors.full_messages.to_sentence
+      redirect_to clone_biomedical_concepts_path(:id => params[:form_id], :namespace => params[:form_namespace])
     end
   end
 
+=begin
   def update
     authorize BiomedicalConcept
     instance = params[:instance]
@@ -98,21 +95,25 @@ class BiomedicalConceptsController < ApplicationController
       render :json => { :errors => @bc.errors.full_messages}, :status => 422
     end
   end
+=end
 
   def destroy
     authorize BiomedicalConcept
-    id = params[:id]
-    namespace = params[:namespace]
-    bc = BiomedicalConcept.find(id, namespace)
-    bc.destroy
-    redirect_to biomedical_concepts_path
+    bc = BiomedicalConcept.find(params[:id], the_params[:namespace])
+    token = Token.obtain(bc, current_user)
+    if !token.nil?
+      bc.destroy
+      AuditTrail.delete_item_event(current_user, form, "Biomedical Concept deleted.")
+      token.release
+    else
+      flash[:error] = "The item is locked for editing by another user."
+    end
+    redirect_to request.referer
   end
 
   def show 
     authorize BiomedicalConcept
-    id = params[:id]
-    namespace = params[:namespace]
-    @bc = BiomedicalConcept.find(id, namespace)
+    @bc = BiomedicalConcept.find(params[:id], the_params[:namespace])
     respond_to do |format|
       format.html do
         @items = @bc.get_properties(true)
@@ -128,33 +129,27 @@ class BiomedicalConceptsController < ApplicationController
 
   def export_ttl
     authorize BiomedicalConcept
-    id = params[:id]
-    namespace = params[:namespace]
-    @bc = IsoManaged::find(id, namespace)
+    bc = IsoManaged.find(params[:id], the_params[:namespace])
     send_data to_turtle(@bc.triples), filename: "#{@bc.owner}_#{@bc.identifier}.ttl", type: 'application/x-turtle', disposition: 'inline'
   end
   
   def export_json
     authorize BiomedicalConcept
-    id = params[:id]
-    namespace = params[:namespace]
-    @bc = BiomedicalConcept.find(id, namespace)
+    bc = IsoManaged.find(params[:id], the_params[:namespace])
     send_data @bc.to_api_json, filename: "#{@bc.owner}_#{@bc.identifier}.json", :type => 'application/json; header=present', disposition: "attachment"
   end
 
   def upgrade
     authorize BiomedicalConcept
-    id = params[:id]
-    namespace = params[:namespace]
-    @bc = BiomedicalConcept.find(id, namespace)
+    @bc = BiomedicalConcept.create(params)
     @bc.upgrade
-    redirect_to history_biomedical_concepts_path(:identifier => @bc.identifier, :scope_id => @bc.owner_id)
+    redirect_to history_biomedical_concept_path(:biomedical_concept => { :identifier => @bc.identifier, :scope_id => @bc.owner_id })
   end
   
 private
 
   def the_params
-    params.require(:biomedical_concept).permit(:data)
+    params.require(:biomedical_concept).permit(:namespace, :uri, :identifier, :label, :scope_id, :bc_id, :bc_namespace, :bct_id, :bct_namespace)
   end  
 
 end
